refactor: centralize auth coredata
