+++ core/templates/site/forgotPasswordNoEmailPage.gohtml
@@ -1,4 +1,7 @@
 {{ template "head" $ }}
+{{- if cd.CurrentError }}
+<p style="color:red">{{ cd.CurrentError }}</p>
+{{- end }}
 <p>No verified email address found for {{ .Username }}.</p>
 <form method="post">
     {{ csrfField }}
