+++ core/templates/site/forgotPasswordPage.gohtml
@@ -1,4 +1,7 @@
 {{ template "head" $ }}
+{{- if cd.CurrentError }}
+<p style="color:red">{{ cd.CurrentError }}</p>
+{{- end }}
 <p>A verification code will be sent to your email address. The new password
 only becomes active after you log in using this code.</p>
 <form method="post">
