+++ core/templates/site/passwordVerifyPage.gohtml
@@ -1,4 +1,7 @@
 {{ template "head" $ }}
+{{- if cd.CurrentError }}
+<p style="color:red">{{ cd.CurrentError }}</p>
+{{- end }}
 <form method="post" action="/login/verify">
     {{ csrfField }}
     <input type="hidden" name="id" value="{{ .ID }}">
