+++ handlers/auth/forgotPassword_no_email_test.go
@@ -24,7 +24,7 @@ func TestForgotPasswordNoEmail(t *testing.T) {
 	}
 	defer conn.Close()
 	q := db.New(conn)
-	mock.ExpectQuery("SystemGetUserByUsername").WillReturnRows(sqlmock.NewRows([]string{"idusers", "email", "username", "public_profile_enabled_at"}).AddRow(1, "", "u", nil))
+	mock.ExpectQuery("SystemGetLogin").WillReturnRows(sqlmock.NewRows([]string{"idusers", "email", "passwd", "passwd_algorithm", "username"}).AddRow(1, "", "", "", "u"))
 	mock.ExpectQuery(regexp.QuoteMeta("SELECT 1 FROM user_roles ur JOIN roles r ON ur.role_id = r.id WHERE ur.users_idusers = ? AND r.can_login = 1 LIMIT 1")).WithArgs(int32(1)).WillReturnRows(sqlmock.NewRows([]string{"column_1"}).AddRow(1))
 
 	cd := common.NewCoreData(context.Background(), q, config.NewRuntimeConfig())
